+++ system-tests/project-fixtures/angular/src/app/components/button-output.component.ts
@@ -2,6 +2,7 @@ import { Component, EventEmitter, Output } from '@angular/core'
 
 @Component({
   selector: 'app-button-output',
+  standalone: false,
   template: `<button (click)="clicked.emit(true)">Click Me</button>`,
 })
 export class ButtonOutputComponent {